refactor(manager): move isLocalhostRequest into middlewares/api
isLocalhostRequest was defined in middlewares/index but its only consumers
are in middlewares/api (csrfGuard and localhostOrAdmin), which imported it
twice via lazy require('./index') to sidestep a load-order cycle. Move the
function to api.js next to its callers: both lazy requires and the cycle
guard disappear, and the duplicate destructure is gone. Removed from the
index barrel's exports (nothing imports it from there); api.js exports it
for discoverability alongside the other guards.
